LPD-31063 Separate Journal autosave tests in a different file
